requirements.txt
# Core dependencies
django==5.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
dj-database-url==2.3.0

# Development tools
pytest-django
pytest-cov

django-deepface==0.0.2
